pyproject.toml
[project]
name = "pencil-bot"
version = "0.0.1"
description = "Telegram GIF bot with keyword detection"
authors = [
    {name = "Pencil Bot", email = "getrisog@gmail.com"}
]
dependencies = [
    "python-telegram-bot==20.7",
    "python-dotenv==1.0.0",
    "pymorphy3",
    "aiohttp==3.9.1",
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[project.scripts]
pencil-bot = "pencil_bot.main:main" 